feat: echo the found PR title (#16)
While running this action, I'd like feedback to confirm that the PR title the action is running matches what I expect. This is especially important when I edit a title and re-run the action; I want to sanity check the action and I want to be able to know what the PR title was in the action history.

## Verifying the change

I wrote a bad PR title and CI reported:

```
Run ./
Found PR title: lolbad title 
Error: PR title: subject may not be empty
Error: PR title: type may not be empty
Error: PULL REQUEST: PR title does not conform to the conventional commit spec
```

Behold: the second line shows the the PR title. And as a bonus, makes it easy to act on the error.